pyproject.toml
[project]
name = "ai-services"
version = "0.1.0"
description = "AI Services for Social Website - Async workers with RabbitMQ and MongoDB"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn==0.23.0",
    "aio-pika==9.4.0",
    "transformers==4.36.0",
    "torch>=2.5.0",           # (Lưu ý 1)
    "sentencepiece==0.1.99",
    "sentence-transformers==2.2.2",
    "motor==3.3.2",           # <-- Chỉ cần cái này
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
